ml_training/requirements.txt
tensorflow>=2.20.0
numpy>=1.24.3
matplotlib>=3.7.2
seaborn>=0.12.2
scikit-learn>=1.3.0
pillow>=10.1.0
pandas>=2.0.3
